+++ runtime/onert/backend/cpu/ops/ArgMinMaxLayer.cc
@@ -17,9 +17,35 @@
 #include "ArgMinMaxLayer.h"
 
 #include "OperationUtils.h"
+#include "../KernelGenerator.h"
+#include "../Validator.h"
 
 #include <cker/operation/ArgMinMax.h>
 #include <assert.h>
+
+namespace onert::backend::cpu
+{
+
+void Validator::visit(const ir::operation::ArgMinMax &) { _supported = true; }
+
+void KernelGenerator::visit(const ir::operation::ArgMinMax &node)
+{
+  const auto output_index{node.getOutputs().at(0)};
+  const auto input_index{node.getInputs().at(ir::operation::ArgMinMax::INPUT)};
+  const auto axis_index{node.getInputs().at(ir::operation::ArgMinMax::AXIS)};
+
+  auto output_tensor = _tensor_reg->getPortableTensor(output_index);
+  auto input_tensor = _tensor_reg->getPortableTensor(input_index);
+  auto axis_tensor = _tensor_reg->getPortableTensor(axis_index);
+
+  auto fn = std::make_unique<ops::ArgMinMaxLayer>();
+
+  fn->configure(input_tensor, output_tensor, axis_tensor, node.param().is_arg_max);
+
+  _return_fn = std::move(fn);
+}
+
+} // namespace onert::backend::cpu
 
 namespace onert::backend::cpu::ops
 {
