+++ runtime/onert/backend/cpu/ops/DynamicUpdateSliceLayer.cc
@@ -15,9 +15,41 @@
  */
 
 #include "DynamicUpdateSliceLayer.h"
+
+#include "../KernelGenerator.h"
+#include "../Validator.h"
+
 #include "OperationUtils.h"
 
 #include <cker/operation/DynamicUpdateSlice.h>
+
+namespace onert::backend::cpu
+{
+
+void Validator::visit(const ir::operation::DynamicUpdateSlice &) { _supported = true; }
+
+void KernelGenerator::visit(const ir::operation::DynamicUpdateSlice &node)
+{
+  using ir::operation::DynamicUpdateSlice;
+
+  const auto output_index{node.getOutputs().at(0)};
+  const auto operand_index{node.getInputs().at(DynamicUpdateSlice::Input::OPERAND)};
+  const auto update_index{node.getInputs().at(DynamicUpdateSlice::Input::UPDATE)};
+  const auto indices_index{node.getInputs().at(DynamicUpdateSlice::Input::INDICES)};
+
+  auto output_tensor = _tensor_reg->getPortableTensor(output_index);
+  auto operand_tensor = _tensor_reg->getPortableTensor(operand_index);
+  auto update_tensor = _tensor_reg->getPortableTensor(update_index);
+  auto indices_tensor = _tensor_reg->getPortableTensor(indices_index);
+
+  auto fn = std::make_unique<ops::DynamicUpdateSliceLayer>();
+
+  fn->configure(operand_tensor, update_tensor, indices_tensor, output_tensor);
+
+  _return_fn = std::move(fn);
+}
+
+} // namespace onert::backend::cpu
 
 namespace onert::backend::cpu::ops
 {
