+++ runtime/onert/backend/cpu/ops/ExpandDimsLayer.cc
@@ -16,6 +16,32 @@
 
 #include "ExpandDimsLayer.h"
 
+#include "../KernelGenerator.h"
+#include "../Validator.h"
+
+namespace onert::backend::cpu
+{
+
+void Validator::visit(const ir::operation::ExpandDims &) { _supported = true; }
+
+void KernelGenerator::visit(const ir::operation::ExpandDims &node)
+{
+  const auto output_index{node.getOutputs().at(0)};
+  const auto input_index{node.getInputs().at(ir::operation::ExpandDims::Input::INPUT)};
+  // AXIS input is used for output shape inference
+
+  auto output_tensor = _tensor_reg->getPortableTensor(output_index);
+  auto input_tensor = _tensor_reg->getPortableTensor(input_index);
+
+  auto fn = std::make_unique<ops::ExpandDimsLayer>();
+
+  fn->configure(input_tensor, output_tensor);
+
+  _return_fn = std::move(fn);
+}
+
+} // namespace onert::backend::cpu
+
 namespace onert::backend::cpu::ops
 {
