+++ runtime/onert/backend/cpu/ops/FillLayer.cc
@@ -17,8 +17,33 @@
 #include "FillLayer.h"
 
 #include "OperationUtils.h"
+#include "../KernelGenerator.h"
+#include "../Validator.h"
 
 #include <cker/operation/Fill.h>
+
+namespace onert::backend::cpu
+{
+
+void Validator::visit(const ir::operation::Fill &) { _supported = true; }
+
+void KernelGenerator::visit(const ir::operation::Fill &node)
+{
+  const auto output_index{node.getOutputs().at(0)};
+  // SHAPE input is used for shape inference
+  const auto value_index{node.getInputs().at(ir::operation::Fill::Input::VALUE)};
+
+  auto output_tensor = _tensor_reg->getPortableTensor(output_index);
+  auto value_tensor = _tensor_reg->getPortableTensor(value_index);
+
+  auto fn = std::make_unique<ops::FillLayer>();
+
+  fn->configure(value_tensor, output_tensor);
+
+  _return_fn = std::move(fn);
+}
+
+} // namespace onert::backend::cpu
 
 namespace onert::backend::cpu::ops
 {
