+++ runtime/onert/backend/cpu/ops/FusedBatchNormLayer.cc
@@ -16,7 +16,37 @@
 
 #include "FusedBatchNormLayer.h"
 
+#include "../KernelGenerator.h"
+#include "../Validator.h"
+
 #include <cker/operation/FusedBatchNorm.h>
+
+namespace onert::backend::cpu
+{
+
+void Validator::visit(const ir::operation::FusedBatchNorm &) { _supported = true; }
+
+void KernelGenerator::visit(const ir::operation::FusedBatchNorm &node)
+{
+  const auto ofm_index{node.getOutputs().at(0)};
+
+  auto output_tensor = _tensor_reg->getPortableTensor(ofm_index);
+  std::vector<const IPortableTensor *> input_tensors;
+  for (const auto &ifm_idx : node.getInputs())
+    input_tensors.emplace_back(_tensor_reg->getPortableTensor(ifm_idx));
+
+  const auto epsilon = node.param().epsilon;
+  const auto is_training = node.param().is_training;
+  const auto &data_format = node.param().data_format;
+
+  auto fn = std::make_unique<ops::FusedBatchNormLayer>();
+
+  fn->configure(input_tensors, epsilon, is_training, data_format, output_tensor);
+
+  _return_fn = std::move(fn);
+}
+
+} // namespace onert::backend::cpu
 
 namespace onert::backend::cpu::ops
 {
