+++ runtime/onert/backend/cpu/ops/GatherLayer.cc
@@ -18,8 +18,37 @@
 
 #include "OperationUtils.h"
 #include "GGMLHelper.h"
+#include "../KernelGenerator.h"
+#include "../Validator.h"
 
 #include <cker/operation/Gather.h>
+
+namespace onert::backend::cpu
+{
+
+void Validator::visit(const ir::operation::Gather &) { _supported = true; }
+
+void KernelGenerator::visit(const ir::operation::Gather &node)
+{
+  const auto output_index{node.getOutputs().at(0)};
+  const auto input_index{node.getInputs().at(ir::operation::Gather::Input::INPUT)};
+  const auto indices_index{node.getInputs().at(ir::operation::Gather::Input::INDICES)};
+
+  auto output_tensor = _tensor_reg->getPortableTensor(output_index);
+  auto input_tensor = _tensor_reg->getPortableTensor(input_index);
+  auto indices_tensor = _tensor_reg->getPortableTensor(indices_index);
+
+  const auto rank = _ctx.at(input_index).shape().rank();
+  const auto axis = ops::getAxis(rank, node.param().axis);
+
+  auto fn = std::make_unique<ops::GatherLayer>();
+
+  fn->configure(input_tensor, indices_tensor, output_tensor, axis, _external_context.get());
+
+  _return_fn = std::move(fn);
+}
+
+} // namespace onert::backend::cpu
 
 namespace onert::backend::cpu::ops
 {
