+++ runtime/onert/backend/cpu/ops/LogSoftmaxLayer.cc
@@ -17,8 +17,35 @@
 #include "LogSoftmaxLayer.h"
 
 #include "OperationUtils.h"
+#include "../KernelGenerator.h"
+#include "../Validator.h"
 
 #include <cker/operation/LogSoftMax.h>
+
+namespace onert::backend::cpu
+{
+
+void Validator::visit(const ir::operation::LogSoftmax &) { _supported = true; }
+
+void KernelGenerator::visit(const ir::operation::LogSoftmax &node)
+{
+  const auto output_index{node.getOutputs().at(0)};
+  const auto input_index{node.getInputs().at(ir::operation::LogSoftmax::Input::INPUT)};
+
+  const auto beta = node.param().beta;
+  const auto axis = node.param().axis;
+
+  auto output_tensor = _tensor_reg->getPortableTensor(output_index);
+  auto input_tensor = _tensor_reg->getPortableTensor(input_index);
+
+  auto fn = std::make_unique<ops::LogSoftMaxLayer>();
+
+  fn->configure(input_tensor, beta, axis, output_tensor);
+
+  _return_fn = std::move(fn);
+}
+
+} // namespace onert::backend::cpu
 
 namespace onert::backend::cpu::ops
 {
