+++ runtime/onert/backend/cpu/ops/ReverseLayer.cc
@@ -17,8 +17,34 @@
 #include "ReverseLayer.h"
 
 #include "OperationUtils.h"
+#include "../KernelGenerator.h"
+#include "../Validator.h"
 
 #include <cker/operation/Reverse.h>
+
+namespace onert::backend::cpu
+{
+
+void Validator::visit(const ir::operation::Reverse &) { _supported = true; }
+
+void KernelGenerator::visit(const ir::operation::Reverse &node)
+{
+  const auto output_index{node.getOutputs().at(0)};
+  const auto input_index{node.getInputs().at(ir::operation::Reverse::INPUT)};
+  const auto axis_index{node.getInputs().at(ir::operation::Reverse::AXIS)};
+
+  auto output_tensor = _tensor_reg->getPortableTensor(output_index);
+  auto input_tensor = _tensor_reg->getPortableTensor(input_index);
+  auto axis_tensor = _tensor_reg->getPortableTensor(axis_index);
+
+  auto fn = std::make_unique<ops::ReverseLayer>();
+
+  fn->configure(input_tensor, axis_tensor, output_tensor);
+
+  _return_fn = std::move(fn);
+}
+
+} // namespace onert::backend::cpu
 
 namespace onert::backend::cpu::ops
 {
