+++ runtime/onert/backend/cpu/ops/TopKV2Layer.cc
@@ -17,8 +17,38 @@
 #include "TopKV2Layer.h"
 
 #include "OperationUtils.h"
+#include "../KernelGenerator.h"
+#include "../Validator.h"
 
 #include <cker/operation/TopKV2.h>
+
+namespace onert::backend::cpu
+{
+
+void Validator::visit(const ir::operation::TopKV2 &) { _supported = true; }
+
+void KernelGenerator::visit(const ir::operation::TopKV2 &node)
+{
+  const auto outputValues_index{node.getOutputs().at(ir::operation::TopKV2::Output::OUTPUT_VALUES)};
+  const auto outputIndices_index{
+    node.getOutputs().at(ir::operation::TopKV2::Output::OUTPUT_INDICES)};
+
+  const auto inputData_index{node.getInputs().at(ir::operation::TopKV2::Input::INPUT)};
+
+  const auto k = node.param().k;
+
+  auto values_tensor = _tensor_reg->getPortableTensor(outputValues_index);
+  auto indices_tensor = _tensor_reg->getPortableTensor(outputIndices_index);
+  auto input_tensor = _tensor_reg->getPortableTensor(inputData_index);
+
+  auto fn = std::make_unique<ops::TopKV2Layer>();
+
+  fn->configure(input_tensor, values_tensor, indices_tensor, k);
+
+  _return_fn = std::move(fn);
+}
+
+} // namespace onert::backend::cpu
 
 namespace onert::backend::cpu::ops
 {
