+++ runtime/onert/backend/cpu/ops/TransposeLayer.cc
@@ -17,9 +17,35 @@
 #include "TransposeLayer.h"
 
 #include "OperationUtils.h"
+#include "../KernelGenerator.h"
+#include "../Validator.h"
 
 #include <cker/operation/Transpose.h>
 #include <numeric>
+
+namespace onert::backend::cpu
+{
+
+void Validator::visit(const ir::operation::Transpose &) { _supported = true; }
+
+void KernelGenerator::visit(const ir::operation::Transpose &node)
+{
+  const auto output_index{node.getOutputs().at(0)};
+  const auto input_index{node.getInputs().at(ir::operation::Transpose::Input::INPUT)};
+  const auto perm_index{node.getInputs().at(ir::operation::Transpose::Input::PERMUTATION)};
+
+  auto output_tensor = _tensor_reg->getPortableTensor(output_index);
+  auto input_tensor = _tensor_reg->getPortableTensor(input_index);
+  auto perm_tensor = _tensor_reg->getPortableTensor(perm_index);
+
+  auto fn = std::make_unique<ops::TransposeLayer>();
+
+  fn->configure(input_tensor, perm_tensor, output_tensor);
+
+  _return_fn = std::move(fn);
+}
+
+} // namespace onert::backend::cpu
 
 namespace onert::backend::cpu::ops
 {
